chore: reorganize internal functions
